Split up __init__ in new package.
